Checkpoint trying to make N-encoder work with attention
Currently discovered a bug where concatenating encoder outputs has a problem when the trace lengths are different.
No problem occurs when the module is ran without attention and all other settings the same.

Dynamic_RNN should produce cell outputs that are padded with zeros to the max timestep possible.
Investigate why this is not the case for attention